+++ infinite-rs/src/lib.rs
@@ -17,7 +17,8 @@ fn load_modules() -> Result<()> {
     // These are the main archive files used in Halo Infinite.
     let mut module = ModuleFile::new();
     // Reads to the module file given a file path.
-    module.read(String::from("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module"))?;
+    // Note: the path can be anything that implements AsRef<Path>.
+    module.read("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module")?;
     Ok(())
 }
 ```
@@ -31,11 +32,9 @@ The `read_tag_from_id` function is also available to load a tag by its global ID
 use infinite_rs::{ModuleFile, Result};
 
 fn load_tags() -> Result<()> {
-    // Create new instance of a Module file.
-    // These are the main archive files used in Halo Infinite.
     let mut module = ModuleFile::new();
-    // Reads to the module file given a file path.
-    module.read(String::from("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module"))?;
+    module.read("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module")?;
+
     // Load a specific tag from the module file.
     let tag_index = 0;
     module.read_tag(tag_index)?;
@@ -49,6 +48,30 @@ fn load_tags() -> Result<()> {
 ## Creating a custom structure and reading it
 `infinite-rs` also allows you to read data directly into structures, using the `read_metadata` function. This functionality requires the `derive` feature.
 
+### Defining Structures
+To define a structure that can be read from a tag data stream, you must first derive the `TagStructure` trait. To ensure proper padding and alignment, you can use the `data` attribute to specify the size of the structure in bytes. Each field also must contain a `data` attribute specifying the offset in bytes from the start of the structure.
+
+> [!TIP]
+> Padding between fields are automatically calculated. Any data between two offsets are skipped.
+
+```rust
+use infinite_rs_derive::TagStructure;
+use infinite_rs::tag::types::common_types::{
+    AnyTag, FieldReference,
+};
+
+#[derive(Default, Debug, TagStructure)]
+#[data(size(0x88))] // Size can be any u64 value.
+struct MaterialTag {
+    #[data(offset(0x00))] // Offset can be any u64 value within the range of the size.
+    any_tag: AnyTag,
+    #[data(offset(0x10))]
+    material_shader: FieldReference,
+}
+```
+
+### Reading structures
+
 ```rust
 use infinite_rs_derive::TagStructure;
 use infinite_rs::tag::types::common_types::{
@@ -57,30 +80,49 @@ use infinite_rs::tag::types::common_types::{
 use infinite_rs::{ModuleFile, Result};
 
 #[derive(Default, Debug, TagStructure)]
-#[data(size(0x88))]
+#[data(size(0x88))] // Size can be any u64 value.
 struct MaterialTag {
-    #[data(offset(0x00))]
+    #[data(offset(0x00))] // Offset can be any u64 value within the range of the size.
     any_tag: AnyTag,
     #[data(offset(0x10))]
     material_shader: FieldReference,
 }
 
 fn load_tags() -> Result<()> {
-    // Create new instance of a Module file.
-    // These are the main archive files used in Halo Infinite.
     let mut module = ModuleFile::new();
-    // Reads to the module file given a file path.
-    module.read(String::from("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module"))?;
-    // Load a specific tag from the module file.
-    let tag_index = 0;
-    module.read_tag(tag_index)?;
-    // We can now read metadata.
-    let tag = &mut module.files[tag_index as usize];
-    let mut mat = MaterialTag::default();
-    tag.read_metadata(&mut mat)?;
+    module.read("C:/XboxGames/Halo Infinite/Content/deploy/any/globals-rtx-new.module")?;
+
+    // We now want to find the material tags in the module file.
+    let material_indices = module.files.iter()
+        .enumerate()
+        .filter(|(_, file)| file.tag_group == "mat ")
+        .map(|(index, _)| index)
+        .collect::<Vec<_>>();
+
+    // And for each material tag, we want to read the metadata associated.
+    for index in material_indices {
+        // We first have to populate data_stream and tag_info.
+        module.read_tag(index as u32)?;
+        let mut mat = MaterialTag::default();
+        // We pass in our structure as a generic parameter.
+        module.files[index].read_metadata(&mut mat)?;
+        // We can now access the fields in our structure.
+        // For instance, `any_tag.internal_struct.tag_id` is always equal to the tag id of our file.
+        assert_eq!(module.files[index].tag_id, mat.any_tag.internal_struct.tag_id);
+    }
     Ok(())
 }
 ```
+
+## Credits
+- [libinfinite](https://github.com/Coreforge/libInfinite) by Coreforge, which this project is mostly based on.
+- [Reclaimer](https://github.com/Gravemind2401/Reclaimer) by Gravemind2401, which helped me get familiar with Blam file formats.
+- [AusarDocs](https://github.com/ElDewrito/AusarDocs) by Shockfire, a very useful resource on Ausar/Slipspace file formats.
+- [Kraken](https://github.com/WolvenKit/kraken) by WolvenKit team, a re-implementation of Oodle Kraken, removing the need for any binary blobs being required for decompression.
+- [TagFramework](https://github.com/Codename-Atriox/TagFramework) by Codename Atriox, which was a common reference point for Slipspace internals.
+- [red4lib](https://github.com/rfuzzo/red4lib) by rfuzzo, acting as the main inspiration for this project.
+- [HIRT](https://github.com/urium1186/HIRT) by urium1186, which was very useful in debugging and verifying output from this project.
+
 */
 
 pub mod common;