+++ infinite-rs/src/module/file.rs
@@ -316,13 +316,14 @@ impl ModuleFileEntry {
             .as_ref()
             .ok_or(Error::TagError(TagError::NoTagInfo))?;
         let main_struct = tag_info
-            .structs
+            .struct_definitions
             .iter()
             .find(|s| s.struct_type == TagStructType::MainStruct)
             .ok_or(Error::TagError(TagError::MainStructNotFound))?;
 
         #[allow(clippy::cast_sign_loss)]
-        let main_block: &TagDataBlock = &tag_info.datablock[main_struct.target_index as usize];
+        let main_block: &TagDataBlock =
+            &tag_info.datablock_definitions[main_struct.target_index as usize];
         let full_tag_buffer = &full_tag[usize::try_from(main_block.offset)?..];
         let mut full_tag_reader = BufReader::new(Cursor::new(full_tag_buffer));
 
@@ -330,8 +331,8 @@ impl ModuleFileEntry {
         struct_type.load_field_blocks(
             main_struct.target_index,
             &mut full_tag_reader,
-            &tag_info.structs[..],
-            &tag_info.datablock[..],
+            &tag_info.struct_definitions[..],
+            &tag_info.datablock_definitions[..],
         )?;
 
         Ok(T::default())